dpll: base framework improvements
1. remove the doit indirection

It is perhaps a good idea overall to use the callback table
and params, but all netlink callbacks are not created equal.
And IDK how to quickly work around that with the table and
params.

2. hoist the replying into dpll_genl_cmd_device_get_id()

Every netlink .doit does not have to reply. Sending empty replies
is probably fine but it serves no purpose.

3. remove the centralized reply handler

4. wrap devices in messages not nests

Standard netlink practice is for every object to be dumped
as a separate msg (note, msg != skb, one skb can hold multiple
struct nlmsghdrs). It makes parsing more natural for user space
and messages have longer length fields than nests.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>