dpll: core: remove device and pin nest on replies
Previously respond to device/pin get commands used a separated
nested attribute for returning a single device/pin object.
Remove that approach, use a new netlink header insted.
Although DPLL_A_DEVICE is still used, but to enclose a single device
when device-get is invoked, only to provide user with devices
a pin is connected with.

Previously respond for dump of device had different format then respond
for `do` request. Now respond to device-get commands of both do and
dump requests are in the same format.

Signed-off-by: Arkadiusz Kubalewski <arkadiusz.kubalewski@intel.com>