dpll: core: fix device_idx and pin_idx to search for records
Rename dev_driver_id to device_idx and pin_idx respectively.
Use those indexes to store a record on pin-dpll references.
Use those indexes to find a record on pin-dpll references.

Global xarray indexes for both dpll_device and dpll_pin structures
are kept in `id` member.
That member is used to erase pin or dpll device as well as to mark the
dpll device.

Signed-off-by: Arkadiusz Kubalewski <arkadiusz.kubalewski@intel.com>