dpll: spec: move pin-direction to pin-dpll tuple
Previously it was possible to have different pin-direction returned to
the user depending on which dpll was given as argument for pin-get cmd.
If pin was registered with multiple dplls each could have reported
different direction. I.e. driver exposes chained pins, where one pin is
an input for one dpll and an output for second dpll.
Fix it my enclosing pin-direction in the 'device' nested attribute.

Signed-off-by: Arkadiusz Kubalewski <arkadiusz.kubalewski@intel.com>